Externalised function to new file
Pushed detect deafen to different file and moved gc dump to message file
